+++ assets/scripts/main.js
@@ -1,6 +1,6 @@
 /** import external dependencies */
 import 'jquery';
-import 'bootstrap/dist/js/bootstrap';
+import 'bootstrap';
 
 /** import local dependencies */
 import Router from './util/Router';