+++ templates/layouts/base.blade.php
@@ -2,15 +2,10 @@
 <html @php(language_attributes())>
   @include('partials.head')
   <body @php(body_class())>
-    <!--[if IE]>
-      <div class="alert alert-warning">
-        {!! __('You are using an <strong>outdated</strong> browser. Please <a href="http://browsehappy.com/">upgrade your browser</a> to improve your experience.', 'sage') !!}
-      </div>
-    <![endif]-->
     @php(do_action('get_header'))
     @include('partials.header')
     <div class="wrap container" role="document">
-      <div class="content row">
+      <div class="content">
         <main class="main">
           @yield('content')
         </main>